% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocusZoomPlot.R
\name{LocusZoomPlot}
\alias{LocusZoomPlot}
\title{LocusZoomPlot}
\usage{
LocusZoomPlot(input_params, width = NULL, height = NULL, elementId = NULL)
}
\description{
R wrapper for the locuszoom.js. This file generates a standard, pre-defined plot from the locuszoom.js library
}
